refactor(mzn_gecode): make functional.
